pyproject.toml
[project]
name = "repl-nix-harmonyanalyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "music21>=9.3.0",
    "reportlab>=4.2.5",
    "pillow>=11.0.0",
    "matplotlib",
    "numpy",
]
